% seq.PCICt trunc.PCICt mean.PCICt range.PCICt rep.PCICt julian.PCICt
%  diff.PCICt cut.PCICt c.PCICt is.numeric.PCICt

\name{c.PCICt}
\alias{rep.PCICt}
\alias{seq.PCICt}
\alias{trunc.PCICt}
\alias{mean.PCICt}
\alias{range.PCICt}
\alias{julian.PCICt}
\alias{diff.PCICt}
\alias{cut.PCICt}
\alias{c.PCICt}
\alias{is.numeric.PCICt}
\alias{c.PCICt}
\alias{min.PCICt}
\alias{max.PCICt}
\title{c.PCICt}
\description{
  These functions implement repetition, cutting, sequences, truncation,
  means, ranges, min, max, julian days, differences, concatenation, and
  tests for numericity.
}

\seealso{
  \code{\link{c}}, \code{\link{rep}}, \code{\link{seq}},
  \code{\link{trunc}}, \code{\link{mean}}, \code{\link{range}},
  \code{\link{julian}}, \code{\link{diff}}, \code{\link{cut}},
  \code{\link{is.numeric}}, \code{\link{max}}, \code{\link{min}}.
}

\keyword{ts}
\keyword{chron}